requirements.txt
Flask==3.1.0
mysql-connector-python==9.3.0
Pillow==12.2.0
pdfkit==1.0.0
python-docx==1.2.0
PyPDF2==3.0.1
